Adicionando redirecionamento de url para novo livro e deletar livro

+++ routes/livro_routes.py
@@ -14,17 +14,17 @@ def novo_livro():
 
     if not titulo or not autor or not quantidade_total:
         flash("Todos os dados são obrigatórios", "error")
-        return redirect(url_for("index"), code=303)
+        return redirect(url_for("index", open="novoLivro"), code=303)
     
     if not quantidade_total.isdigit():
         flash("Quantidade deve conter apenas números", "error")
-        return redirect(url_for("index"), code=303)
+        return redirect(url_for("index", open="novoLivro"), code=303)
     
     try:
         quantidade_total = int(quantidade_total)
     except (TypeError, ValueError):
         flash("Quantidade deve ser um número!", "error")
-        return redirect(url_for("index"), code=303)
+        return redirect(url_for("index", open="novoLivro"), code=303)
     
     livro = Livro(
         titulo=titulo,
@@ -40,10 +40,10 @@ def novo_livro():
         db.session.rollback()
 
         flash("Esse Livro já foi cadastrado.", "error")
-        return redirect(url_for("index"), code=303)
+        return redirect(url_for("index", open="novoLivro"), code=303)
     
     flash("Livro criado com sucesso!", "success")
-    return redirect(url_for("index"), code=303)
+    return redirect(url_for("index", open="novoLivro"), code=303)
 
 @livro_bp.route("/deletar", methods=["POST"])
 def deletar_livro():
@@ -51,21 +51,17 @@ def deletar_livro():
 
     if not id:
         flash("ID do Livro é obrigatório", "error")
-        return redirect(url_for("index"), code=303)
+        return redirect(url_for("index", open="deletarLivro"), code=303)
     
     try:
         id = int(id)
     except (TypeError, ValueError):
         flash("ID deve ser um número!", "error")
-        return redirect(url_for("index"), code=303)
-    
-    if not id.isdigit():
-        flash("ID tem que ser um número!", "error")
-        return redirect(url_for("index"), code=303)
+        return redirect(url_for("index", open="deletarLivro"), code=303)
 
     if id <= 0:
         flash("ID inválido", "error")
-        return redirect(url_for("index"), code=303)
+        return redirect(url_for("index", open="deletarLivro"), code=303)
     
     # antes
     #livro = Livro.query.get(id)
@@ -73,15 +69,15 @@ def deletar_livro():
     #depois
     livro = db.session.get(Livro, id)
 
-    if not livro:
+    if not livro or livro is None:
         flash("Livro não encontrado no sistema", "error")
-        redirect(url_for("index"), code=303)
+        return redirect(url_for("index", open="deletarLivro"), code=303)
     
     db.session.delete(livro)
     db.session.commit()
 
     flash("Livro deletado com sucesso!", "success")
-    redirect(url_for("index"), code=303)
+    return redirect(url_for("index", open="deletarLivro"), code=303)
 
 @livro_bp.route("/listar", methods=["GET"])
 def listar_livros():
